refactor(wire): type-safety fixes from external gnitz-wire review
Implements the 3 of 9 external-review findings that validated as
unambiguous improvements; the other 6 were skipped (already mitigated,
lateral, or large blast radius).

- german_string: guard encode_german_string against >u32::MAX lengths
  with a release-build assert instead of silently truncating the 4-byte
  length field of a persisted format.
- WireSysCol.type_code: u8 -> typed TypeCode. No wire/disk format change
  (TypeCode::U64 as u8 == type_code::U64). Consumers updated: gnitz-core
  schema_from_wire_cols drops a from_validated_u8 hop; engine from_wire_cols,
  CONTROL_SCHEMA_DESC, and the wire_sys_schema test use const-stable as-u8.
- decode_op_node: replace the swap-prone (u64,u16,u64,u64) tuple with a
  named CircuitNodeColumn { kind, position, value1, value2 }. Updates the
  decode body, both callers (core from_rows, engine load_circuit), and
  wire tests. Encode path and the GatherReduce legacy path stay on tuples.

Verified: build + clippy --all-targets clean; gnitz-wire/core unit tests,
make test (engine, 1087), and make e2e (895 passed) all green.

+++ crates/gnitz-engine/src/catalog/sys_tables.rs
@@ -199,7 +199,7 @@ const fn from_wire_cols(cols: &[gnitz_wire::WireSysCol], pk_indices: &[u32]) -> 
     let mut buf = [zero_col(); crate::schema::MAX_COLUMNS];
     let mut i = 0;
     while i < cols.len() {
-        buf[i] = SchemaColumn::new(cols[i].type_code, if cols[i].nullable { 1 } else { 0 });
+        buf[i] = SchemaColumn::new(cols[i].type_code as u8, if cols[i].nullable { 1 } else { 0 });
         i += 1;
     }
     let (head, _) = buf.split_at(cols.len());
